pyproject.toml
[project]
name = "boone-gifts-backend"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.5",
    "uvicorn>=0.40.0",
    "sqlalchemy>=2.0.46",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymysql>=1.1.2",
    "cryptography>=46.0.0",
    "pyjwt>=2.11.0",
    "bcrypt>=5.0.0",
    "alembic>=1.18.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "httpx>=0.28.1",
]
